Make testing functions reusable
